added nextgen likeexpr
